Implement/robot log into commands (#99)
* Added a log for storage sub and intake sub

* Add basic log topics

* Added Command Loggers folder and classes

Added a loggers folder in the command folder, and added a logger class for each subsystem.

* Moved CommandLoggers to their own folder

* Move loggers into their own folder, and refactor them for more generality

* Preparing to delete

* Delete everything and start over

This isn't going the way I want it to, so I'm starting over

* Create RobotLog.java

Create and begin putting classes into RobotLog.java

* Add base for every subsystem to RobotLog class

* Start over (again)

Once again, I have failed.

* Create CommandLoggerBase.java

Create a class that extends off of CommandBase. All commands should extend off of this class instead of command base. (If they need to extend off of command group or something then they should do that)

* Refactor commands to use CommandLoggerBase

* Make CommandLoggerBase an abstract class

* Remove all super.execute()

execute logging was not needed, so I removed it.

* Create method to log status to command topic

* Remove non-related files from patch

Co-authored-by: Eli Monnig <35583627+Stick-2357@users.noreply.github.com>